* build/ac-macros/svn-macros.m4 (SVN_CONFIG_NICE):     Stop leaving a config.nice.old, as requested by jerenkrantz.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@860706 13f79535-47bb-0310-9956-ffa450edef68

+++ build/ac-macros/svn-macros.m4
@@ -9,7 +9,11 @@
 # invocation, instead of the entire configured state.
 AC_DEFUN([SVN_CONFIG_NICE], [
   AC_MSG_NOTICE([creating $1])
-  mv "$1" "$1.old"
+  # This little dance satisfies Cygwin, which cannot overwrite in-use files.
+  if test -f "$1"; then
+    mv "$1" "$1.old"
+    rm -f "$1.old"
+  fi
 
   cat >"$1" <<EOF
 #! /bin/sh
